Add test run with testcontainers
