+++ labs/unicorn-store/software/unicorn-store-spring/pom.xml
@@ -7,7 +7,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>3.1.3</version>
+        <version>3.2.1</version>
         <relativePath/>
         <!-- lookup parent from repository -->
     </parent>
@@ -230,11 +230,22 @@
                     <plugin>
                         <groupId>org.springframework.boot</groupId>
                         <artifactId>spring-boot-maven-plugin</artifactId>
-                        <configuration>
+                        <executions>
+                            <execution>
+                                <id>repackage</id>
+                                <configuration>
+                                    <classifier>exec</classifier>
+                                    <mainClass>com.unicorn.store.StoreApplication</mainClass>
+                                    <layout>JAR</layout>
+                                </configuration>
+                            </execution>
+                        </executions>
+                        <!-- <configuration>
+                            <id>repackage</id>
                             <classifier>exec</classifier>
                             <mainClass>com.unicorn.store.StoreApplication</mainClass>
                             <layout>JAR</layout>
-                        </configuration>
+                        </configuration> -->
                     </plugin>
                     <plugin>
                         <groupId>com.google.cloud.tools</groupId>
