--- labs/unicorn-store/software/unicorn-store-spring/src/test/java/TestApplication.java
@@ -1,11 +1,0 @@
-import org.springframework.boot.SpringApplication;
-import com.unicorn.store.StoreApplication;
-
-public class TestApplication {
-    public static void main(String[] args) {
-        SpringApplication
-                .from(StoreApplication::main)
-                .with(ContainersConfig.class)
-                .run(args);
-    }
-}
