+++ labs/unicorn-store/software/unicorn-store-spring/src/test/java/com/unicorn/store/ContainersConfig.java
@@ -1,11 +1,10 @@
+package com.unicorn.store;
+
 import org.springframework.boot.devtools.restart.RestartScope;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
 import org.springframework.context.annotation.Bean;
 import org.testcontainers.containers.PostgreSQLContainer;
-
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
 
 @TestConfiguration(proxyBeanMethods = false)
 public class ContainersConfig {
@@ -15,8 +14,9 @@ public class ContainersConfig {
     @RestartScope
     PostgreSQLContainer<?> postgreSQLContainer(){
         return new PostgreSQLContainer<>("postgres:13.11")
-            .withDatabaseName("unicorns")
-            .withUsername("postgres")
-            .withPassword("postgres");
+        .withDatabaseName("unicorns")
+        .withUsername("postgres")
+        .withPassword("postgres");
     }
+
 }
