+++ labs/unicorn-store/software/unicorn-store-spring/src/test/java/com/unicorn/store/TestApplication.java
@@ -1,0 +1,12 @@
+package com.unicorn.store;
+
+import org.springframework.boot.SpringApplication;
+
+public class TestApplication {
+    public static void main(String[] args) {
+        SpringApplication
+            .from(StoreApplication::main)
+            .with(ContainersConfig.class)
+            .run(args);
+    }
+}
