+++ labs/unicorn-store/software/unicorn-store-spring/src/test/java/com/unicorn/store/UnicornControllerTest.java
@@ -20,6 +20,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
+import org.springframework.test.context.TestPropertySource;
 
 import com.unicorn.store.model.Unicorn;
